Extend csv output
WIP

+++ testCases/casual/001.single.service.scale.payload.py
@@ -1,10 +1,17 @@
-from locust import FastHttpUser, task, events
+from locust import FastHttpUser, task, events, constant
 import base64
 import requests
 import os
-from helpers import *
+from helpers import resetCasualMetrics, getCasualMetrics, writeStartInformation, writeStopInformation
+import time
+import csv
 
-global_host = "http://host2:8201"
+starttime=float()
+
+@events.init_command_line_parser.add_listener
+def _(parser):
+    parser.add_argument("--use-remote-nodes", default=False, action='store_true', help="Remote node mode")
+
 global_1B = base64.b64encode( bytes( 1))
 global_1K = base64.b64encode( bytes( 1 * 1024))
 global_10K = base64.b64encode( bytes( 10 * 1024))
@@ -12,12 +19,14 @@ global_100K = base64.b64encode( bytes( 100 * 1024))
 
 class TestCase( FastHttpUser):
    """ Single service scale payload """
+   wait_time = constant(1)
+
 
    @task
    def task1( self):
       self.client.post(
          name = "1B",
-         path = "/casual/casual/example/echo",
+         url = "/casual/casual/example/echo",
          headers = { "content-type": "application/casual-x-octet"},
          data = global_1B)
 
@@ -25,7 +34,7 @@ class TestCase( FastHttpUser):
    def task2( self):
       self.client.post(
          name = "1K",
-         path = "/casual/casual/example/echo",
+         url = "/casual/casual/example/echo",
          headers = { "content-type": "application/casual-x-octet"},
          data = global_1K)
 
@@ -33,7 +42,7 @@ class TestCase( FastHttpUser):
    def task3( self):
       self.client.post(
          name = "10K",
-         path = "/casual/casual/example/echo",
+         url = "/casual/casual/example/echo",
          headers = { "content-type": "application/casual-x-octet"},
          data = global_10K)
 
@@ -41,32 +50,39 @@ class TestCase( FastHttpUser):
    def task4( self):
       self.client.post(
          name = "100K",
-         path = "/casual/casual/example/echo",
+         url = "/casual/casual/example/echo",
          headers = { "content-type": "application/casual-x-octet"},
          data = global_100K)
 
-
 @events.test_start.add_listener
 def on_test_start( environment, **kwargs):
+   global starttime
    # Warm up
    for i in range( 10):
       requests.post(
-         url = global_host + "/casual/casual/example/echo",
+         url = environment.host + "/casual/casual/example/echo",
          headers = { "content-type": "application/casual-x-octet"},
          data = global_1K)
 
    # Start telegraf
-   os.system( "ssh cas201@host2 'nohup telegraf -config telegraf.conf &>/dev/null &'")
-   
+   if environment.parsed_options.use_remote_nodes:
+      os.system( "ssh cas201@host2 'nohup telegraf -config telegraf.conf &>/dev/null &'")
+
+   starttime = writeStartInformation( environment)
+
    # Reset casual metrics
-   resetCasualMetrics( global_host)
+   resetCasualMetrics( environment.host)
 
 @events.test_stop.add_listener
 def on_test_stop( environment, **kwargs):
-   # Stop telegraf
-   os.system( "ssh cas201@host2 kill $(ssh cas201@host2 ps | grep telegraf | awk '{print $1}')")
-   os.system( "scp cas201@host2:metrics/telegraf.txt $HOME/testResults/001.cas201.telegraf.metrics.txt")
-   
-   # Get casual metrics
-   getCasualMetrics( global_host, "001", "cas201")
+   global starttime
 
+   # Stop telegraf
+   if environment.parsed_options.use_remote_nodes:
+      os.system( "ssh cas201@host2 kill $(ssh cas201@host2 ps | grep telegraf | awk '{print $1}')")
+      os.system( "scp cas201@host2:metrics/telegraf.txt $HOME/testResults/001.cas201.telegraf.metrics.txt")
+
+   # Get casual metrics
+   getCasualMetrics( environment.host, "001", "cas201")
+
+   writeStopInformation( environment, starttime)
